EventsSearchBE/bot-requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
httpx==0.25.2
pandas==2.1.3
sqlalchemy==2.0.23
databases==0.8.0
mysql-connector-python==8.2.0
PyMySQL==1.1.0